removed Admin MVC, finished RESTful path to Order Resource

+++ app/Controller/OrdersController.php
@@ -1,30 +1,55 @@
 <?php
 App::uses('AppController', 'Controller');
+App::uses('RestUtilities', 'Lib');
 /**
  * Orders Controller
  *
  * @property Order $Order
  */
 class OrdersController extends AppController {
-	protected $allowedOrderFields = array(
-		'name',
-		'email',
-		'comments',
-		'payment_kind', 
-		'payment_key', 
-		'payment_status', 
-		'payment_description', 
-		'payment_amount',
-		'payment_env'
+	protected $allowedCreateFields = array(
+		'Order' => array(
+			'name',
+			'email',
+			'comments',
+			'payment_kind', 
+			'payment_key', 
+			'payment_status', 
+			'payment_description', 
+			'payment_amount',
+			'payment_currency',
+			'payment_env'
+		)
 	);
 
+	protected $allowedUpdateFields = array(
+		'Order' => array(
+			'name',
+			'email',
+			'comments',
+			'token',
+			'payment_kind', 
+			'payment_key', 
+			'payment_status', 
+			'payment_description', 
+			'payment_amount',
+			'payment_currency',
+			'payment_env'
+		)
+	);	
+	
 	protected $paginate = array(
 		'limit' => 10
 	);
 	
+	protected $RestUtilities;
+	
 	public function beforeFilter () {        
         parent::beforeFilter();
-		$this->Security->unlockedActions = array('add');
+		if ($this->isRest()) {
+			$this->Security->unlockedActions = array($this->params['action']);
+		}
+		$this->RestUtilities = new RestUtilities();
     }
 
 /**
@@ -36,7 +61,7 @@ class OrdersController extends AppController {
 		if($this->isRest()){
 			$data = array();
 			$offset = isset($this->request->query['offset']) ? $this->request->query['offset'] : '1';
-			$limit = isset($this->request->query['limit']) ? $this->request->query['limit'] : '10';
+			$limit = isset($this->request->query['limit']) ? $this->request->query['limit'] : '100';
 			
 			if(!preg_match('/^[1-9][0-9]*$/',$offset) || !preg_match('/^[1-9][0-9]*$/',$limit)) {
 				$msg = sprintf(__('Please check if "offset" or "limit" are set as numeric numbers from 1 to 999999'));
@@ -51,6 +76,7 @@ class OrdersController extends AppController {
 			$data['offset'] = $offset;
 			$data['limit'] = $limit;
 			
+			$data['status'] = "ok";
 			$this->set(compact('data'));
 		} else {
 			$this->Order->recursive = 0;
@@ -70,8 +96,11 @@ class OrdersController extends AppController {
 			if (!$this->Order->exists($id)) {
 				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid order')));
 			}
+			$data = array();
 			$options = array('conditions' => array('Order.' . $this->Order->primaryKey => $id));
 			$data = $this->Order->find('first', $options);
+			
+			$data['status'] = "ok";
 			$this->set(compact('data'));
 		} else {	
 			if (!$this->Order->exists($id)) {
@@ -92,24 +121,20 @@ class OrdersController extends AppController {
 			if (!$this->request->is('post')) {
 				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid request')));
 			}
+			$data = array();
+			$request_data = $this->RestUtilities->filterFieldsOfRequest($this->request->data, $this->allowedCreateFields);
 			
-			$keys = array_keys($this->request->data['Order']);
-			foreach($keys as $k) {
-				if(!array_search($k, $allowedOrderFields)){
-					unset($this->request->data['Order'][$k]);
-				}
-			}
-
-			if ($this->Order->save($this->request->data)) {
+			if ($this->Order->save($request_data)) {
 				$options = array('conditions' => array('Order.' . $this->Order->primaryKey => $this->Order->id));
-				$order = $this->Order->find('first', $options);
-				
+				$order = $this->Order->find('first', $options);				
 				$data['message'] = 'The order has been saved';
 				$data['order_id'] = $order['Order']['id'];
-				$data['order_token'] = $order['Order']['token'];
+				$data['order_token'] = $order['Order']['token'];				
 			} else {
 				return $this->Rest->abort(array('status' => '400', 'error' => __('The order could not be saved. Please, try again.')));
-			}			
+			}
+			
+			$data['status'] = "ok";
 			$this->set(compact('data'));			
 		} else if ($this->request->is('post')) {
 			$this->Order->create();
@@ -130,20 +155,47 @@ class OrdersController extends AppController {
  * @return void
  */
 	public function edit($id = null) {
-	
-		if (!$this->Order->exists($id)) {
-			throw new NotFoundException(__('Invalid order'));
-		}
-		if ($this->request->is('post') || $this->request->is('put')) {
-			if ($this->Order->save($this->request->data)) {
-				$this->Session->setFlash(__('The order has been saved'), 'flash/success');
-				$this->redirect(array('action' => 'index'));
-			} else {
-				$this->Session->setFlash(__('The order could not be saved. Please, try again.'), 'flash/error');
+		if($this->isRest()){
+			if (!$this->request->is('put') && !$this->request->is('post')) {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid request')));
 			}
-		} else {
+			$data = array();			
+			if (!$this->Order->exists($id)) {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid request, record not found')));
+			}
+			
 			$options = array('conditions' => array('Order.' . $this->Order->primaryKey => $id));
-			$this->request->data = $this->Order->find('first', $options);
+			$order = $this->Order->find('first', $options);			
+			
+			if(!isset($this->request->data['Order']['token']) || strcmp($order['Order']['token'],$this->request->data['Order']['token'])!=0) {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('Invalid token')));
+			}
+			
+			$request_data = $this->RestUtilities->filterFieldsOfRequest($this->request->data, $this->allowedUpdateFields);
+			$request_data['Order']['id'] = $id;
+			
+			if ($this->Order->save($request_data)) {				
+				$data['message'] = 'The order has been updated';
+			} else {
+				return $this->Rest->abort(array('status' => '400', 'error' => __('The order could not be saved. Please, try again.')));
+			}
+			$data['status'] = "ok";
+			$this->set(compact('data'));
+		} else {
+			if (!$this->Order->exists($id)) {
+				throw new NotFoundException(__('Invalid order'));
+			}
+			if ($this->request->is('post') || $this->request->is('put')) {
+				if ($this->Order->save($this->request->data)) {
+					$this->Session->setFlash(__('The order has been saved'), 'flash/success');
+					$this->redirect(array('action' => 'index'));
+				} else {
+					$this->Session->setFlash(__('The order could not be saved. Please, try again.'), 'flash/error');
+				}
+			} else {
+				$options = array('conditions' => array('Order.' . $this->Order->primaryKey => $id));
+				$this->request->data = $this->Order->find('first', $options);
+			}
 		}
 	}
 
@@ -156,16 +208,24 @@ class OrdersController extends AppController {
  * @return void
  */
 	public function delete($id = null) {
-		$this->Order->id = $id;
-		if (!$this->Order->exists()) {
-			throw new NotFoundException(__('Invalid order'));
-		}
-		$this->request->onlyAllow('post', 'delete');
-		if ($this->Order->delete()) {
-			$this->Session->setFlash(__('Order deleted'), 'flash/success');
+		if($this->isRest()){
+			$data = array();
+			
+			$data['message'] = __('Method not supported');
+			$data['status'] = "fail";
+			$this->set(compact('data'));
+		} else {		
+			$this->Order->id = $id;
+			if (!$this->Order->exists()) {
+				throw new NotFoundException(__('Invalid order'));
+			}
+			$this->request->onlyAllow('post', 'delete');
+			if ($this->Order->delete()) {
+				$this->Session->setFlash(__('Order deleted'), 'flash/success');
+				$this->redirect(array('action' => 'index'));
+			}
+			$this->Session->setFlash(__('Order was not deleted'), 'flash/error');
 			$this->redirect(array('action' => 'index'));
 		}
-		$this->Session->setFlash(__('Order was not deleted'), 'flash/error');
-		$this->redirect(array('action' => 'index'));
 	}
 }
